check if mfn is supported by IOCTL_PRIVCMD_MMAPBATCH before calling ioctl()
If mfn is valid but is large enough to have the top nibble (the
"error nibble") set then there is no way to do proper error
reporting, because the mfn and the error code will get intermixed.
So we should check the input mfn before calling ioctl(). If the
user doesn't want a mapping at a given index, he can use ~0UL.

Signed-off-by: Wen Congyang <wency@cn.fujitsu.com>